+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 * ![Feature][badge-feature] Allow users to inject custom JavaScript resources to enable alternatives to Google Analytics like plausible.io. ([#1706][github-1706])
 
+* ![Feature][badge-feature] The keyword argument `strict` in `makedocs` is more flexible: in addition to a boolean indicating whether or not any error should result in a failure, `strict` also accepts a `Symbol` or `Vector{Symbol}` indicating which error(s) should result in a build failure. ([1689][github-1689])
+
 ## Version `v0.27.7`
 
 * ![Bugfix][badge-bugfix] Fix an error when building documentation for the first time with `push_preview`. ([#1693][github-1693], [#1704][github-1704])
@@ -925,6 +927,7 @@
 [github-1673]: https://github.com/JuliaDocs/Documenter.jl/pull/1673
 [github-1687]: https://github.com/JuliaDocs/Documenter.jl/pull/1687
 [github-1691]: https://github.com/JuliaDocs/Documenter.jl/pull/1691
+[github-1689]: https://github.com/JuliaDocs/Documenter.jl/pull/1689
 [github-1693]: https://github.com/JuliaDocs/Documenter.jl/issues/1693
 [github-1696]: https://github.com/JuliaDocs/Documenter.jl/pull/1696
 [github-1698]: https://github.com/JuliaDocs/Documenter.jl/issues/1698